Feat: 반 삭제 로직 추가
