+++ src/main/java/com/edubill/edubillApi/service/auth/AuthServiceImpl.java
@@ -1,4 +1,4 @@
-package com.edubill.edubillApi.service;
+package com.edubill.edubillApi.service.auth;
 
 import com.edubill.edubillApi.domain.AuthInfo;
 import com.edubill.edubillApi.domain.User;
@@ -6,7 +6,7 @@ import com.edubill.edubillApi.domain.enums.AuthRole;
 import com.edubill.edubillApi.dto.user.LoginRequestDto;
 import com.edubill.edubillApi.dto.user.SignupRequestDto;
 import com.edubill.edubillApi.dto.user.UserDto;
-import com.edubill.edubillApi.dto.verification.VerificationResponseDto;
+import com.edubill.edubillApi.dto.auth.VerificationResponseDto;
 import com.edubill.edubillApi.error.exception.UserAlreadyExistsException;
 import com.edubill.edubillApi.error.exception.UserNotFoundException;
 import com.edubill.edubillApi.repository.redis.RequestIdRepository;
