+++ src/test/java/com/edubill/edubillApi/controller/AuthControllerTest.java
@@ -3,16 +3,16 @@ package com.edubill.edubillApi.controller;
 import com.edubill.edubillApi.config.TestcontainerConfig;
 import com.edubill.edubillApi.dto.user.ExistUserRequestDto;
 
-import com.edubill.edubillApi.dto.verification.VerificationRequestDto;
-import com.edubill.edubillApi.dto.verification.VerificationResponseDto;
+import com.edubill.edubillApi.dto.auth.VerificationRequestDto;
+import com.edubill.edubillApi.dto.auth.VerificationResponseDto;
 import com.edubill.edubillApi.jwt.JwtProvider;
 import com.edubill.edubillApi.repository.redis.RequestIdRepository;
 import com.edubill.edubillApi.repository.redis.RequestIdRepositoryMap;
 import com.edubill.edubillApi.repository.users.UserRepository;
 import com.edubill.edubillApi.repository.verification.VerificationRepository;
 import com.edubill.edubillApi.repository.verification.VerificationRepositoryMap;
-import com.edubill.edubillApi.service.AuthService;
-import com.edubill.edubillApi.service.AuthServiceMock;
+import com.edubill.edubillApi.service.auth.AuthService;
+import com.edubill.edubillApi.service.auth.AuthServiceMock;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validator;
