+++ src/test/java/com/edubill/edubillApi/excel/ExcelServiceTest.java
@@ -1,11 +1,9 @@
 package com.edubill.edubillApi.excel;
 
 import com.edubill.edubillApi.config.TestcontainerConfig;
-import com.edubill.edubillApi.dto.ExcelResponseDto;
+import com.edubill.edubillApi.service.excel.ExcelService;
 import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -21,8 +19,6 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.time.YearMonth;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeoutException;
 
 
 @Slf4j
